backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incube-backend"
version = "1.0.0"
description = "InCube Platform — AI-powered business transformation API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36,<2.1.0",
    "asyncpg>=0.30.0,<0.32.0",
    "alembic>=1.14.0,<2.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "redis>=5.2.0,<8.0.0",
    "miniopy-async>=1.21",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "anthropic>=0.42.0",
    "reportlab>=4.1",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "respx>=0.22.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
